command_line: Remove duplicate output for battery status
The list controller.battery_voltage_control_register_names
and the method controller.get_battery_voltage_control_registers()
are actually duplicates to the following methods.

While it is fine to have both ways of reading the
battery_voltage_control_registers from a library perspective (and we
definitely want to have this feature for writing), we do not need to
output all values two times on the command line.

Removing the variant with the least pretty formatting.